move Aggregation to datastructures folder
