+++ SOMcreator/datastructure/__init__.py
@@ -2,3 +2,4 @@ from .project import Project
 from .object_ import Object
 from .propertyset import PropertySet
 from .attribute import Attribute
+from .aggregation import Aggregation
